Migrate session init to thunk
