+++ src/platform/plugins/shared/discover/public/application/main/components/top_nav/on_save_search.test.tsx
@@ -14,26 +14,14 @@ import { dataViewMock } from '@kbn/discover-utils/src/__mocks__';
 import { dataViewWithTimefieldMock } from '../../../../__mocks__/data_view_with_timefield';
 import { onSaveSearch } from './on_save_search';
 import { savedSearchMock } from '../../../../__mocks__/saved_search';
-import { getDiscoverStateContainer } from '../../state_management/discover_state';
 import type { ReactElement } from 'react';
 import { discoverServiceMock } from '../../../../__mocks__/services';
 import type { SavedSearch } from '@kbn/saved-search-plugin/public';
-import { createBrowserHistory } from 'history';
-import { mockCustomizationContext } from '../../../../customizations/__mocks__/customization_context';
-import { createInternalStateStore, createRuntimeStateManager } from '../../state_management/redux';
-import type { HistoryLocationState } from '../../../../build_services';
+import { getDiscoverStateMock } from '../../../../__mocks__/discover_state.mock';
 
 function getStateContainer({ dataView }: { dataView?: DataView } = {}) {
   const savedSearch = savedSearchMock;
-  const history = createBrowserHistory<HistoryLocationState>();
-  const services = { ...discoverServiceMock, history };
-  const runtimeStateManager = createRuntimeStateManager();
-  const stateContainer = getDiscoverStateContainer({
-    services,
-    customizationContext: mockCustomizationContext,
-    internalState: createInternalStateStore({ services, runtimeStateManager }),
-    runtimeStateManager,
-  });
+  const stateContainer = getDiscoverStateMock({});
   stateContainer.savedSearchState.set(savedSearch);
   stateContainer.appState.getState = jest.fn(() => ({
     rowsPerPage: 250,
