--- src/platform/plugins/shared/discover/public/application/main/hooks/use_discover_state_container.ts
@@ -1,40 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the "Elastic License
- * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
- * Public License v 1"; you may not use this file except in compliance with, at
- * your election, the "Elastic License 2.0", the "GNU Affero General Public
- * License v3.0 only", or the "Server Side Public License, v 1".
- */
-
-import { useState } from 'react';
-import type {
-  DiscoverStateContainer,
-  DiscoverStateContainerParams,
-} from '../state_management/discover_state';
-import { getDiscoverStateContainer } from '../state_management/discover_state';
-import { createInternalStateStore } from '../state_management/redux';
-
-/**
- * Creates a state container using the initial params and allows to reset it.
- * The container is recreated only when reset is called. This is useful to reset Discover to its initial state.
- * @param params
- */
-export const useDiscoverStateContainer = (
-  params: Omit<DiscoverStateContainerParams, 'internalState'>
-) => {
-  const [stateContainer, setStateContainer] = useState<DiscoverStateContainer>(() =>
-    getDiscoverStateContainer({ ...params, internalState: createInternalStateStore(params) })
-  );
-
-  return [
-    stateContainer,
-    {
-      reset: () => {
-        setStateContainer(
-          getDiscoverStateContainer({ ...params, internalState: createInternalStateStore(params) })
-        );
-      },
-    },
-  ] as const;
-};
