+++ src/platform/plugins/shared/discover/public/application/main/state_management/redux/actions/index.ts
@@ -8,3 +8,4 @@
  */
 
 export * from './data_views';
+export * from './initialize_session';
