+++ src/platform/plugins/shared/discover/public/application/main/state_management/redux/index.ts
@@ -11,6 +11,7 @@ import { omit } from 'lodash';
 import { internalStateSlice, loadDataViewList } from './internal_state';
 import {
   appendAdHocDataViews,
+  initializeSession,
   replaceAdHocDataViewWithId,
   setAdHocDataViews,
   setDataView,
@@ -29,6 +30,7 @@ export const internalStateActions = {
   setDefaultProfileAdHocDataViews,
   appendAdHocDataViews,
   replaceAdHocDataViewWithId,
+  initializeSession,
 };
 
 export {
